src/main/java/org/failearly/dataz/template/encoder/SimpleEncoder.java
/*
 * dataZ - Test Support For Data Stores.
 *
 * Copyright (C) 2014-2016 'Marko Umek' (http://fail-early.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>
 */

package org.failearly.dataz.template.encoder;

import org.failearly.common.annotations.Tests;
import org.failearly.dataz.config.Constants;
import org.failearly.dataz.internal.template.encoder.SimpleEncoderFactory;
import org.failearly.dataz.template.Scope;
import org.failearly.dataz.template.TemplateObjectFactory;

import java.lang.annotation.*;

/**
 * SimpleEncoder is a Template Object Annotation and is responsible for ... .
 */
@Target({ElementType.METHOD, ElementType.TYPE})
@Retention(RetentionPolicy.RUNTIME)
@Documented
@Repeatable(SimpleEncoder.List.class)
@TemplateObjectFactory.Definition(SimpleEncoderFactory.class)
@Tests("SimpleEncoderTest")
public @interface SimpleEncoder {
    /**
     * @return The name of the template object. Could be used in Velocity templates by {@code $<name>}.
     */
    String name();

    /**
     * @return The name of the associated dataz.
     */
    String dataset() default Constants.DATASET_DEFAULT_NAME;

    /**
     * @return The scope of the template object (either {@link Scope#LOCAL} or {@link Scope#GLOBAL}.
     */
    Scope scope() default Scope.DEFAULT;

    /**
     * @return the simple encoder type.
     */
    Type type();


    /**
     * Used by @Repeatable.
     */
    @Target({ElementType.METHOD, ElementType.TYPE})
    @Retention(RetentionPolicy.RUNTIME)
    @Documented
    @interface List {
        SimpleEncoder[] value();
    }

    /**
     * The supported simple encoders.
     */
    enum Type {
        /**
         * No encoding
         */
        NONE,
        /**
         * String to Hex
         */
        HEX,
        /**
         * String to base64
         */
        BASE64
    }
}

